toujours aussi nul

+++ headers/list_utils.h
@@ -17,7 +17,7 @@
 # include <limits.h>
 
 void	free_stack(t_stack	*stack);
-int		lst_size(t_stack *stack);
+size_t	lst_size(t_stack *stack);
 void	free_both(t_stack **a, t_stack **b);
 void	put_index(t_stack **stack);
 t_stack	*lst_last(t_stack *stack);
